move LeftContent into its own component
