+++ src/components/UploadDocumentCard/styles.tsx
@@ -1,15 +1,13 @@
 // External Imports
 // React
-import React from 'react';
 // Material
 import { css, LinearProgress, styled } from '@mui/material';
 
 // Relative Imports
 // Components
 import { Flex } from '../Flex';
-import { Typography } from '../Typography';
 // Styling
-import { UploadDocumentCardState, SharedProps } from './index';
+import { UploadDocumentCardState } from './index';
 
 const FlexCard = styled(Flex)<{ state: UploadDocumentCardState }>`
   width: 442px;
@@ -49,24 +47,4 @@ const StyledLinearProgress = styled(LinearProgress)(({ theme }) => ({
   },
 }));
 
-type LeftContentProps = {
-  state: UploadDocumentCardState;
-} & Pick<SharedProps, 'fileName' | 'fileSize'>;
-
-const LeftContent: React.FC<LeftContentProps> = ({ state, fileName, fileSize }) => {
-  const isError = state === 'error';
-  return (
-    <Flex direction="column" grow="1">
-      <Typography class="medium" size="sm" padding={0.4} color={isError ? 'error75' : 'dark75'}>
-        {fileName ?? 'Select a file or drag and drop here'}
-      </Typography>
-      {state !== 'locked' && (
-        <Typography class="roman" size="xs" padding={0.4} color={isError ? 'error60' : 'dark60'}>
-          {fileSize ?? 'maximum file size is 200mb'}
-        </Typography>
-      )}
-    </Flex>
-  );
-};
-
-export { FlexCard, StyledLinearProgress, LeftContent };
+export { FlexCard, StyledLinearProgress };
